Refactor: Remove duplicate SearXNG installation logic
